+++ tapehash/cli.py
@@ -4,14 +4,17 @@ from .work import work, check_difficulty, calculate_difficulty
 
 
 def help_tapehash1():
-    print('Use: command | tapehash1 OR tapehash1 --preimage {str} OR tapehash1 --file path/to/file')
+    print(
+        'Use: command | tapehash1 OR tapehash1 --preimage {str} OR tapehash1 '
+        '--file path/to/file'
+    )
     print('\t--from:hex - parse preimage as hexadecimal')
-    print('\t--code_size {int} - set the code_size parameter (default=1024)')
-    print('\t-cs {int} - set the code_size parameter (default=1024)')
+    print('\t--code_size {int} - set the code_size parameter (default=20)')
+    print('\t-cs {int} - set the code_size parameter (default=20)')
     print('\t--to:raw - write the raw hash digest bytes to stdout')
     print('\t--difficulty - output the difficulty of the hash digest')
-    print('\t--check {int} - outputs 1 if the hash digest meets the ' +
-        'difficulty threshold; outputs 0 and exits with error code 2 '+
+    print('\t--check {int} - outputs 1 if the hash digest meets the '
+        'difficulty threshold; outputs 0 and exits with error code 2 '
         'if it does not'
     )
     print('\nOnly 1 of --to:raw --difficulty --check can be used at a time')
@@ -85,16 +88,19 @@ def run_tapehash1() -> None:
     print(tapehash1(preimage, code_size).hex())
 
 def help_tapehash2():
-    print('Use: command | tapehash1 OR tapehash1 --preimage {str} OR tapehash2 --file path/to/file')
-    print('\t--from:hex - parse preimage as hexadecimal')
-    print('\t--tape_size_multiplier {int} - set the tape_size_multiplier' +
-        ' parameter (default=1024)'
+    print(
+        'Use: command | tapehash2 OR tapehash2 --preimage {str} OR tapehash2 '
+        '--file path/to/file'
     )
-    print('\t-tsm {int} - set the tape_size_multiplier parameter (default=1024)')
+    print('\t--from:hex - parse preimage as hexadecimal')
+    print('\t--tape_size_multiplier {int} - set the tape_size_multiplier '
+        'parameter (default=2)'
+    )
+    print('\t-tsm {int} - set the tape_size_multiplier parameter (default=2)')
     print('\t--to:raw - write the raw hash digest bytes to stdout')
     print('\t--difficulty - output the difficulty of the hash digest')
-    print('\t--check {int} - outputs 1 if the hash digest meets the ' +
-        'difficulty threshold; outputs 0 and exits with error code 2 '+
+    print('\t--check {int} - outputs 1 if the hash digest meets the '
+        'difficulty threshold; outputs 0 and exits with error code 2 '
         'if it does not'
     )
     print('\nOnly 1 of --to:raw --difficulty --check can be used at a time')
@@ -168,18 +174,21 @@ def run_tapehash2() -> None:
     print(tapehash2(preimage, tape_size_multiplier).hex())
 
 def help_tapehash3():
-    print('Use: command | tapehash1 OR tapehash1 --preimage {str} OR tapehash3 --file path/to/file')
-    print('\t--from:hex - parse preimage as hexadecimal')
-    print('\t--code_size {int} - set the code_size parameter (default=1024)')
-    print('\t-cs {int} - set the code_size parameter (default=1024)')
-    print('\t--tape_size_multiplier {int} - set the tape_size_multiplier' +
-        ' parameter (default=1024)'
+    print(
+        'Use: command | tapehash3 OR tapehash3 --preimage {str} OR tapehash3 '
+        '--file path/to/file'
     )
-    print('\t-tsm {int} - set the tape_size_multiplier parameter (default=1024)')
+    print('\t--from:hex - parse preimage as hexadecimal')
+    print('\t--code_size {int} - set the code_size parameter (default=64)')
+    print('\t-cs {int} - set the code_size parameter (default=64)')
+    print('\t--tape_size_multiplier {int} - set the tape_size_multiplier' +
+        ' parameter (default=2)'
+    )
+    print('\t-tsm {int} - set the tape_size_multiplier parameter (default=2)')
     print('\t--to:raw - write the raw hash digest bytes to stdout')
     print('\t--difficulty - output the difficulty of the hash digest')
-    print('\t--check {int} - outputs 1 if the hash digest meets the ' +
-        'difficulty threshold; outputs 0 and exits with error code 2 '+
+    print('\t--check {int} - outputs 1 if the hash digest meets the '
+        'difficulty threshold; outputs 0 and exits with error code 2 '
         'if it does not'
     )
     print('\nOnly 1 of --to:raw --difficulty --check can be used at a time')
@@ -195,7 +204,7 @@ def run_tapehash3() -> None:
 
     if len(argv) > 1:
         if '--help' in argv or '-h' in argv:
-            return help_tapehash2()
+            return help_tapehash3()
 
         if '--preimage' in argv:
             idx = argv.index('--preimage')
@@ -244,11 +253,15 @@ def run_tapehash3() -> None:
             tape_size_multiplier = int(argv[idx+1])
 
         if '--to:raw' in argv:
-            stdout.buffer.write(tapehash3(preimage, code_size, tape_size_multiplier))
+            stdout.buffer.write(tapehash3(
+                preimage, code_size, tape_size_multiplier
+            ))
             return
 
         if '--difficulty' in argv:
-            diff = calculate_difficulty(tapehash3(preimage, code_size, tape_size_multiplier))
+            diff = calculate_difficulty(tapehash3(
+                preimage, code_size, tape_size_multiplier
+            ))
             print(diff)
             return
 
@@ -258,7 +271,7 @@ def run_tapehash3() -> None:
                 print("must supply integer after --check")
                 exit(1)
             target = int(argv[idx+1])
-            digest = tapehash1(preimage, code_size, tape_size_multiplier)
+            digest = tapehash3(preimage, code_size, tape_size_multiplier)
             if check_difficulty(digest, target):
                 print(1)
                 return